requirements.txt
ipykernel==6.29.4
openai==1.35.3
python-dotenv==1.0.1